sql rules
