+++ test/Dapper.AOT.Test/Verifiers/DAP004.cs
@@ -10,46 +10,46 @@ public class DAP004 : Verifier<WrappedDapperInterceptorAnalyzer>
 {
     [Fact]
     public Task LanguageTooLow() => CSVerifyAsync("""
-using Dapper;
-using System.Data.Common;
+        using Dapper;
+        using System.Data.Common;
 
-[DapperAot(true)]
-class SomeCode
-{
-    public void Foo(DbConnection conn) => conn.Execute("some sql");
-}
-""", new[]
+        [DapperAot(true)]
+        class SomeCode
+        {
+            public void Foo(DbConnection conn) => conn.Execute("some sql");
+        }
+        """, new[]
     {
         InterceptorsEnabled, WithCSharpLanguageVersion(LanguageVersion.CSharp10)
     }, Diagnostic(DapperInterceptorGenerator.Diagnostics.LanguageVersionTooLow));
 
     [Fact]
     public Task CSFineIfInactive() => CSVerifyAsync("""
-using Dapper;
-using System.Data.Common;
+        using Dapper;
+        using System.Data.Common;
 
-[DapperAot(false)]
-class SomeCode
-{
-    public void Foo(DbConnection conn) => conn.Execute("some sql");
-}
-""", new[]
+        [DapperAot(false)]
+        class SomeCode
+        {
+            public void Foo(DbConnection conn) => conn.Execute("some sql");
+        }
+        """, new[]
     {
         InterceptorsEnabled, WithCSharpLanguageVersion(LanguageVersion.CSharp10)
     });
 
     [Fact]
     public Task VBFineIfInactive() => VBVerifyAsync("""
-Imports Dapper
-Imports System.Data.Common
+        Imports Dapper
+        Imports System.Data.Common
 
-' [DapperAot(false)]
-Module SomeCode
-    Public Sub Foo(conn As DbConnection)
-        conn.Execute("some sql")
-    End Sub
-End Module
-""", new[]
+        ' [DapperAot(false)]
+        Module SomeCode
+            Public Sub Foo(conn As DbConnection)
+                conn.Execute("some sql")
+            End Sub
+        End Module
+        """, new[]
 {
         InterceptorsEnabled, WithCSharpLanguageVersion(LanguageVersion.CSharp10)
     });
