+++ test/Dapper.AOT.Test/Verifiers/DAP005.cs
@@ -9,47 +9,47 @@ public class DAP005 : Verifier<WrappedDapperInterceptorAnalyzer>
 {
     [Fact]
     public Task ShouldFlagWhenUsedAndNotAttrib() => CSVerifyAsync("""
-using Dapper;
-using System.Data.Common;
+        using Dapper;
+        using System.Data.Common;
 
-class SomeCode
-{
-    public void Foo(DbConnection conn) => conn.Execute("some sql");
-}
-""", Diagnostic(DapperInterceptorGenerator.Diagnostics.DapperAotNotEnabled));
+        class SomeCode
+        {
+            public void Foo(DbConnection conn) => conn.Execute("some sql");
+        }
+        """, Diagnostic(DapperInterceptorGenerator.Diagnostics.DapperAotNotEnabled));
 
     [Fact]
     public Task ShouldNotFlagWhenNotUsedAndNoAttrib() => CSVerifyAsync("""
-using Dapper;
-using System.Data.Common;
+        using Dapper;
+        using System.Data.Common;
 
-class SomeCode
-{
-    public void Foo(DbConnection conn) {}
-}
-""");
+        class SomeCode
+        {
+            public void Foo(DbConnection conn) {}
+        }
+        """);
 
     [Fact]
     public Task ShouldNotFlagWhenUsedAndOptedOut() => CSVerifyAsync("""
-using Dapper;
-using System.Data.Common;
+        using Dapper;
+        using System.Data.Common;
 
-[DapperAot(false)]
-class SomeCode
-{
-    public void Foo(DbConnection conn) => conn.Execute("some sql");
-}
-""");
+        [DapperAot(false)]
+        class SomeCode
+        {
+            public void Foo(DbConnection conn) => conn.Execute("some sql");
+        }
+        """);
 
     [Fact]
     public Task ShouldNotFlagWhenUsedAndOptedIn() => CSVerifyAsync("""
-using Dapper;
-using System.Data.Common;
+        using Dapper;
+        using System.Data.Common;
 
-[DapperAot(true)]
-class SomeCode
-{
-    public void Foo(DbConnection conn) => conn.Execute("some sql");
-}
-""", InterceptorsGenerated(1, 1, 1, 0, 0));
+        [DapperAot(true)]
+        class SomeCode
+        {
+            public void Foo(DbConnection conn) => conn.Execute("some sql");
+        }
+        """, InterceptorsGenerated(1, 1, 1, 0, 0));
 }
