+++ test/Dapper.AOT.Test/Verifiers/DAP009.cs
@@ -1,23 +1,24 @@
-﻿using Dapper.CodeAnalysis;
+﻿using Dapper.AOT.Test.TestCommon;
+using Dapper.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using System.Threading.Tasks;
 using Xunit;
 
 namespace Dapper.AOT.Test.Verifiers;
 
-public class DAP009 : Verifier<DapperInterceptorGenerator>
+public class DAP009 : Verifier<WrappedDapperInterceptorAnalyzer>
 {
     [Fact]
     public Task LanguagesTooLow() => CSVerifyAsync("""
-using Dapper;
-using System.Data.Common;
+        using Dapper;
+        using System.Data.Common;
 
-[DapperAot(true)]
-class SomeCode
-{
-    public void Foo(DbConnection conn) => conn.Execute("some sql");
-}
-""", new[]
+        [DapperAot(true)]
+        class SomeCode
+        {
+            public void Foo(DbConnection conn) => conn.Execute("some sql");
+        }
+        """, new[]
     {
         InterceptorsEnabled, WithCSharpLanguageVersion(LanguageVersion.CSharp10)
     }, Diagnostic(DapperInterceptorGenerator.Diagnostics.LanguageVersionTooLow));
